move strings -> strings-advanced and add new strings exercise
* move strings to strings-advanced; new strings story for formatting

strings exercise (creating and formatting strings)

* add design.md and adjust design.md for strings-advanced

* remove test

* add config.json

* requested fixes

* add party robot story to reference/types/string.md

* use prettier

* differentiate between strings and strings-advanced in prerequisites

* Update languages/exercises/concept/strings-advanced/.docs/instructions.md


* Update languages/exercises/concept/strings-advanced/.docs/instructions.md


* rename strings-advanced to strings-package exercise

* changes from review
