+++ exercises/concept/strings-package/.docs/hints.md
