+++ exercises/concept/strings-package/.meta/config.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "authors": [
     {
       "github_username": "tehsphinx",
