+++ exercises/concept/strings-package/.meta/design.md
@@ -1,0 +1,25 @@
+## Goal
+
+The goal of this exercise is to teach the student about the strings package.
+
+## Learning objectives
+
+- Know about the `strings` package (like `Split`, `Index`, `TrimSpace`).
+
+## Out of scope
+
+- iteration of strings
+- immutability of strings
+- bytes and runes
+
+## Concepts
+
+- `strings`
+
+## Prequisites
+
+- `slices`
+
+## Analyzer
+
+See https://github.com/exercism/v3/issues/355.
