+++ exercises/concept/strings-package/go.mod
@@ -1,0 +1,3 @@
+module strings_advanced
+
+go 1.13
