+++ exercises/concept/strings-package/strings_test.go
@@ -1,4 +1,4 @@
-package strings
+package strings_package
 
 import (
 	"strings"
