+++ reference/implementing-a-concept-exercise.md
@@ -89,7 +89,7 @@ If you have any questions regarding implementing the exercise, please post them 
 [docs-rationale-for-v3]: ../../../docs/rationale-for-v3.md
 [docs-features-of-v3]: ../../../docs/features-of-v3.md
 [anatomy-of-a-concept-exercise]: https://www.youtube.com/watch?v=gkbBqd7hPrA
-[concept-exercise-basic-strings]: ../exercises/concept/strings
+[concept-exercise-basic-strings]: ../exercises/concept/strings-package
 [concept-exercise-numbers]: ../exercises/concept/numbers
 [reference]: ../../../reference
 [style guide]: ../docs/style-guide.md
